requirements.txt
django==1.11.15
python-etcd==0.4.5
futures==3.2.0
kubernetes==6.0.0
pika==0.12.0
ansible-tower-cli==3.3.0
dwebsocket==0.4.2
openshift==0.7.0
python-etcd==0.4.5

